Moved functions to db index
